Working Game No Hud
Will work on fixing the hud

+++ Source/SpaceInvader/Public/GameController.h
@@ -27,7 +27,7 @@ public:
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "My Variables")
 	TArray<int> WaveDifficulty;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "My Variables")
-	TArray<int> WaveFrequency;
+	TArray<float> WaveFrequency;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "My Variables")
 	float xPos;
